requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
python-dateutil==2.8.2
gunicorn==21.2.0
packaging==23.2
requests==2.31.0
flasgger==0.9.7.1
